pyproject.toml
[tool.poetry]
name = "course-paper-7"
version = "0.1.0"
description = ""
authors = ["Nikolay Averin <rizzzon116@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
ipython = "^8.26.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
django-phonenumber-field = "^8.0.0"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
